Revert "remove useless condition"
Actually this condition was not useless. A forced move would be flagged dangerous, and wouldn't be pruned. Without that we can return an uninitialized move at low depth in case the only legal move has a negative SEE (SEE pruning).

Refresh the compiles with bug fix.

No functional change (unless in the rare buggy cases)
signature = 7252380