rook pair penalty Contrary to the Bishop pair, the Rook pair should have a disadvantage, as shown by many Q vs. 2R games where the Q proved stronger. A classical interpretation is redundancy. Following the same logic, N are redundant and Q are redundant with R and with themselves as well as with B, although these effects are unlikely to be measurable.
A key theme in lost tournament games is the overestimation of 2R vs Q. We need a specific degree of freedom, because the value of the R has other side effects (makes exchanging a rook for a  minor + 2 pawns more likely for instance).

SPRT 10"+0.1": alpha=beta=0.1 elo0=0 elo1=5. Testing resolution was probably too big, and SPRT was going for a long refutation of the patch. Anyway, I ran out of patience:
Score of redundancy vs master: 3448 - 3357 - 5455  [0.504] 12260

So this patch should probably not be commited, due to the added complexity. Will try to remove is and decrease the value of a rook equivalently now.

signature = 6775453