TT pruning at PV nodes only when ply >= 2 To ensure that: - we always set the best move in search (when best_score raised at the root). - we always have a ponder move pv[1]
This is a slight functional patch, although it does not change the signature. In order to make the functional change apparent, one needs to re-search the same position twice, without flushing the TT in between.

Not doing TT pruning at all for PV nodes, was an "unmeasurable regression":
10000 games in 5"+0.05" with Hash=1: 2839-2875-4286 [49.8%]

So limit to ply >= 2 instead, as suggested by H.G. Muller.